% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_opt.R
\name{CleanMolSlot}
\alias{CleanMolSlot}
\title{Clean Up Molecules Slot}
\usage{
CleanMolSlot(obj)
}
\arguments{
\item{obj}{Seurat object (spatial)}
}
\value{
a Seurat object
}
\description{
This function is removes any molecules that are not found in the FOV, essentially reducing the size of the Seurat object.
}
