% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cells_in_polygon.R
\name{get_cells_in_polygon}
\alias{get_cells_in_polygon}
\title{Get Cells in Polygon}
\usage{
get_cells_in_polygon(seurat.obj, poly_df, image_name)
}
\arguments{
\item{seurat.obj}{Seurat object with spatial data}

\item{poly_df}{Data frame of polygon coordinates (x & y as columns)}

\item{image_name}{Name of image to look for cells in}
}
\value{
List of data frames
}
\description{
This function will take a polygon data frame (x & y coordinates) and parse out which cells are in that polygon.
}
