% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_polygons.R
\name{parse_polygons}
\alias{parse_polygons}
\title{Parse Polygons}
\usage{
parse_polygons(df)
}
\arguments{
\item{df}{Data frame with polygone data}
}
\value{
List of data frames
}
\description{
This function will parse a data frame where the geometry column is formated like this: POLYGON ((x1 y1, x2 y2)). It will out put a list of data frames.
}
