% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLR.R
\name{PlotLR}
\alias{PlotLR}
\title{Ligand-Receptor plotting - mean of mean expressions}
\usage{
PlotLR(
  results,
  ligands = NULL,
  receptors = NULL,
  l.cells = NULL,
  r.cells = NULL,
  split.subset = NULL,
  split.along.y = FALSE,
  use.adj.pval = FALSE,
  pval.threshold = 0.05,
  min.exp.percent = 0.1,
  min.cell.percent = 0,
  resample = 1000
)
}
\arguments{
\item{results}{data.frame containing output from \code{\link{StandardLR}}.}

\item{ligands}{Character vector with specific ligands of interest to plot. If
\code{NULL}, all available ligand genes will be plotted.}

\item{receptors}{Character vector with sepcific receptors of interest to
plot. If \code{NULL}, all available receptor genes will be plotted.}

\item{l.cells}{Character vector with specific cell clusters to plot as the
ligand-expressing cell-type. If \code{NULL}, all available cell cluster types
will be plotted.}

\item{r.cells}{Character vector with specific cell clusters to plot as the
receptor-expressing cell-type. If \code{NULL}, all available cell cluster
types will be plotted.}

\item{split.subset}{Character vector with values of \code{split.by} column
from \code{results} to retain. All other values will be removed from
plotting. If \code{NULL}, all available values of \code{split.by} will be
plotted.}

\item{split.along.y}{Logical whether \code{split.by} plot facets should run
horizontally (\code{FALSE}) or vertically (\code{TRUE}).}

\item{use.adj.pval}{Logical whether to plot adjusted p-values column from
\code{results}. Default is \code{FALSE}.}

\item{pval.threshold}{Numeric value of the maximum p-value or adjusted
p-value to plot. All scores with p-values greater than the threshold will be
removed.}

\item{min.exp.percent}{Numeric value of the minimum percent that a cell
cluster must express a gene in order to be plotted. Percent expresses below
this value will not be plotted.}

\item{min.cell.percent}{Numeric value (need to expand more).}

\item{resample}{Numeric value of the number of resamplings used in the
permutation test.}
}
\value{
A ggplot2 object with the ligand-receptor plot.
}
\description{
Ligand-Receptor plotting - mean of mean expressions
}
