% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetImageBoundary.R
\name{SetImageBoundary}
\alias{SetImageBoundary}
\title{Set Boundary of Images}
\usage{
SetImageBoundary(seurat_obj, boundary)
}
\arguments{
\item{seurat_obj}{Seurat object (spatial)}

\item{boundary}{Boundary name (e.g. "segmentation" or "centroids")}
}
\value{
a Seurat object
}
\description{
This function changes the default boundary of a spatial Seurat object with multiple images.
}
