% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_opt.R
\name{subset_opt}
\alias{subset_opt}
\title{Subset A Spatial Seurat Object}
\usage{
subset_opt(
  object = NULL,
  subset = NULL,
  cells = NULL,
  idents = NULL,
  features = NULL,
  Update.slots = TRUE,
  Update.object = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A dataframe of with 4 iterations of filtering (Keep vs. Filter)
}
\description{
This function is an updated version of subset to properly subset Cosmx
Seurat objects. It has the same parameters as subset.
}
