% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cell_cycle_phase.R
\name{assign_cell_cycle_phase}
\alias{assign_cell_cycle_phase}
\title{Determine Cell Cycle Phase}
\usage{
assign_cell_cycle_phase(
  seurat_obj,
  s_genes,
  g2m_genes,
  threshold_quantile = 0.5
)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{s_genes}{S Phase Genes (must be species specific)}

\item{g2m_genes}{G2M Phase Genes (must be species specific)}

\item{threshold_quantile}{Threshold for G1 Phase, 0.5 is the recommended threshold}
}
\value{
Vector containing child GO terms
}
\description{
This function acts the same as Seurat's CellCycleScoring, but it uses AddModuleScore_UCell
}
