% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fovs.R
\name{combine_fovs}
\alias{combine_fovs}
\title{Combine FOVs}
\usage{
combine_fovs(
  obj,
  assay = "RNA",
  n_cols = 2,
  offset = 5000,
  fov_name = "combined",
  append = TRUE
)
}
\arguments{
\item{obj}{Seurat object (spatial)}

\item{assay}{GEX Assay}

\item{n_cols}{Number of columns for combined FOVs}

\item{offset}{Number of pixels for offset}

\item{fov_name}{Name of combined FOV}
}
\value{
a Seurat object
}
\description{
This function combines multiple FOVs into one using a predetermined offset
}
