% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAndIntegrateRNA.R
\name{CreateAndIntegrateRNA}
\alias{CreateAndIntegrateRNA}
\title{Create and Integrate Seurat RNA Objects}
\usage{
CreateAndIntegrateRNA(
  data_dirs,
  cells = 3,
  features = 200,
  treatment = NULL,
  use_quantile = TRUE,
  quantile_value_min = 0.15,
  feature_min = NA,
  feature_max = NA,
  percent_mt_max = NA,
  interactive = FALSE,
  object_names = NULL,
  cell_IDs = names(seurat_objects),
  to_regress = "percent.mt",
  cluster_resolution = 0.3,
  max_dims = 15,
  use_SCT = TRUE,
  save_rds_file = TRUE,
  file_name = NULL,
  use_elbow_plot = FALSE,
  spatial = FALSE,
  integration = "HarmonyIntegration",
  integration_normalization = "SCT",
  integration_assay = "SCT",
  integration_reduction = "pca",
  new_reduction = "harmony",
  k_anchor = NULL,
  k_weight = NULL
)
}
\arguments{
\item{data_dirs}{Path to the directories}
}
\value{
An integrated Seurat object
}
\description{
This function creates multiple Seurat objects and then integrates them. It
takes a list of directories as input. In each directory, there should be
the contents of the filtered feature matrix folder, a filtered feature
matrix folder, or a .h5 file. You can directly give the output folder from
cellranger into this function. It will preferentially choose the
filtered_feature_matrix folder. The function then automatically filters each
objects and integrates using Harmony. The user can change these parameters.
This function also save multiple Seurat objects along the way and
runs FindAllMarkers on the clusters.
}
