% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateATACObjectsFilter.R
\name{CreateATACObjectsFilter}
\alias{CreateATACObjectsFilter}
\title{Create and Filter Seurat ATAC Objects}
\usage{
CreateATACObjectsFilter(
  data_dirs,
  treatment = NULL,
  filter = TRUE,
  interactive = FALSE,
  peak_region_fragments_min = 3000,
  peak_region_fragments_max = 1e+05,
  pct_reads_in_peaks_min = 40,
  blacklist_ratio_max = 0.025,
  nucleosome_signal_max = 4,
  TSS.enrichment_min = 2,
  peak_region_max = 3000,
  peakwidths_max = 10000,
  peakwidths_min = 20,
  passed_filters_value = 500
)
}
\arguments{
\item{data_dirs}{Path to the directories}
}
\value{
A list of filtered Seurat objects
}
\description{
This function creates multiple Seurat objects and filters them. It takes a
list of directories as input. In each directory, there should be at least the
following files: peaks.bed, singlecell.csv and fragments.tsv.gz. While
reading in the data, the function creates a common peak list from the
samples. Following this, peaks on scaffolds and not the major chromosomes are
removed. You can directly give the output folder from cellranger into this
function. The objects are then automatically filtered based on hard cutoffs.
This can function can be run interactively to choose cutoffs without having
to re-run it.
}
