% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRNAObjects.R
\name{CreateRNAObjects}
\alias{CreateRNAObjects}
\title{Create Seurat RNA Objects}
\usage{
CreateRNAObjects(
  data_dirs,
  cells = 3,
  features = 200,
  treatment = NULL,
  object_names = NULL
)
}
\arguments{
\item{data_dirs}{Path to the directories}
}
\value{
A list of Seurat objects
}
\description{
This function creates multiple Seurat objects. It takes a list of directories
as input. In each directory, there should be the contents of the filtered
feature matrix folder, a filtered feature matrix folder, or a .h5 file. You
can directly give the output folder from cellranger into this function.
It will preferentially choose the filtered_feature_matrix folder.
}
