% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRNAObjects.R
\name{CreateRNAObjects}
\alias{CreateRNAObjects}
\title{Create Seurat RNA Objects}
\usage{
CreateRNAObjects(
  data_dirs,
  cells = 3,
  features = 200,
  treatment = NULL,
  object_names = NULL
)
}
\arguments{
\item{data_dirs}{Path to directories containing matrix.mtx, features.tsv, and
barcodes.tsv or .h5 files.}

\item{cells}{Features must be expressed in at least this many cells}

\item{features}{Cells must have at least this many features}

\item{treatment}{Treatment metadata column (e.g., Age, chemical, etc.)}

\item{object_names}{Names for the Seurat objects}
}
\value{
A list of Seurat objects
}
\description{
This function creates multiple Seurat objects. It takes a list of directories
as input. In each directory, there should be the contents of the filtered
feature matrix folder, a filtered feature matrix folder, or a .h5 file. You
can directly give the output folder from cellranger into this function.
It will preferentially choose the filtered_feature_matrix folder.
}
