% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRNAObjectsFilter.R
\name{CreateRNAObjectsFilter}
\alias{CreateRNAObjectsFilter}
\title{Create and Filter Seurat RNA Objects}
\usage{
CreateRNAObjectsFilter(
  data_dirs,
  cells = 3,
  features = 200,
  treatment = NULL,
  use_quantile = TRUE,
  quantile_value_min = 0.15,
  feature_min = NA,
  feature_max = NA,
  percent_mt_max = NA,
  interactive = FALSE
)
}
\arguments{
\item{data_dirs}{Path to the directories}
}
\value{
A list of filtered Seurat objects
}
\description{
This function creates multiple Seurat objects and filters them. It takes a
list of directories as input. In each directory, there should be the contents
of the filtered feature matrix folder, a filtered feature matrix folder, or
a .h5 file. You can directly give the output folder from cellranger into this
function. It will preferentially choose the filtered_feature_matrix folder.
The objects are then automatically filtered based on the 15th
and 85th percentile for nFeature_RNA and percent.mt. This can function can
be run interactively to choose cutoffs without having to re-run it.
}
