% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVisiumObjects.R
\name{CreateVisiumObjects}
\alias{CreateVisiumObjects}
\title{Create Visium Spatial Objects}
\usage{
CreateVisiumObjects(
  data_dirs,
  treatment = NULL,
  object_names = NULL,
  file_type = "h5"
)
}
\arguments{
\item{data_dirs}{Path to the data directories}

\item{treatment}{Treatment variable for each sample}

\item{object_names}{Names of each object}

\item{file_type}{File type (e.g., h5 or directory)}
}
\value{
A list of Seurat Spatial objects
}
\description{
This function takes creates Seurat objects of Visium data, of various
formats to make it easier if you download data from GEO
}
