% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeDetectionVisium.R
\name{EdgeDetectionVisium}
\alias{EdgeDetectionVisium}
\title{Identify Spots at Edges in Visium Data}
\usage{
EdgeDetectionVisium(
  coord_path,
  seurat.obj = NULL,
  search = "radius",
  neighbors = 7
)
}
\arguments{
\item{coord_path}{Path to the directory with tissue coordinate file}

\item{seurat.obj}{Seurat object that corresponds to coord_path to ensure
barcodes are the same}

\item{search}{Search type (see nn2)}

\item{neighbors}{Number of neighbors to look for}
}
\value{
A dataframe of with 4 iterations of filtering (Keep vs. Filter)
}
\description{
This function takes coordinate data for all spots in the Visium data and
determines which are at the edges of the capture area, tissue boundary,
and tears within the tissue.
}
