% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_children.R
\name{get_all_children}
\alias{get_all_children}
\title{Recursive function to get all child terms for a given GO term}
\usage{
get_all_children(go_term)
}
\arguments{
\item{go_term}{Parent GO term}
}
\value{
Vector containing child GO terms
}
\description{
This function will get all child GO terms
}
