% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_coords.R
\name{get_all_coords}
\alias{get_all_coords}
\title{Get coordinates from all FOVs}
\usage{
get_all_coords(seurat_obj, meta_cols = NULL)
}
\arguments{
\item{obj}{Seurat object (spatial)}
}
\value{
a data frame
}
\description{
This function collect the coordinates from each FOV and combine them into a single data frame
}
