% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadXenium2.R
\name{LoadXenium2}
\alias{LoadXenium2}
\title{Read and Load 10X Genomics Xenium Updated}
\usage{
LoadXenium2(
  data_dir,
  sample_name,
  outs = c("matrix", "microns"),
  type = c("centroids", "segmentations"),
  mols.qv.threshold = 20
)
}
\arguments{
\item{data_dir}{Path to the directory containing all Xenium data}

\item{sample_name}{Name of Sample}

\item{outs}{Molecular outputs to read
\itemize{
\item "matrix": counts matrix
\item "microns": molecule coordinates
}}

\item{type}{Cell spatial coordinate matrices to read
\itemize{
\item "centroids": cell centroids
\item "segmentations": cell segmentations
}}

\item{mols.qv.threshold}{Remvoe transcript molecules wiht a QV less than
specified threshold (20 is recommended)}
}
\value{
A list of filtered Seurat objects
}
\description{
This function is an updated version of LoadXenium from Seurat, as the
original function cannot use the new output of Xenium.
}
