% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSeurat.R
\name{MergeSeurat}
\alias{MergeSeurat}
\title{Merge Seurat RNA Objects}
\usage{
MergeSeurat(
  seurat_objects,
  cell_IDs = names(seurat_objects),
  to_regress = "percent.mt",
  cluster_resolution = 0.3,
  max_dims = 15,
  use_SCT = TRUE,
  save_rds_file = TRUE,
  file_name = NULL,
  use_elbow_plot = FALSE,
  spatial = FALSE,
  integration = "HarmonyIntegration",
  integration_normalization = "SCT",
  integration_assay = "SCT",
  integration_reduction = "pca",
  new_reduction = "harmony",
  k_anchor = NULL,
  k_weight = NULL
)
}
\arguments{
\item{seurat_objects}{List of Seurat objects}
}
\value{
An integrated Seurat object.
}
\description{
This function merges and integrates a list of Seurat objects. By default, it
assumes the objects are SCT normalized and will be integrated using Harmony.
It can integrate ATAC objects, as well. This function also save multiple
Seurat objects along the way and runs FindAllMarkers on the clusters.
}
