% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSeurat.R
\name{MergeSeurat}
\alias{MergeSeurat}
\title{Merge Seurat RNA Objects}
\usage{
MergeSeurat(
  seurat_objects,
  cell_IDs = names(seurat_objects),
  to_regress = "percent.mt",
  cluster_resolution = 0.3,
  max_dims = 15,
  use_SCT = TRUE,
  sct_assay = "RNA",
  save_rds_file = TRUE,
  file_name = NULL,
  use_elbow_plot = FALSE,
  spatial = "no",
  integration = "HarmonyIntegration",
  integration_normalization = "SCT",
  integration_assay = "SCT",
  integration_reduction = "pca",
  new_reduction = "harmony",
  k_anchor = NULL,
  k_weight = NULL,
  markers = TRUE
)
}
\arguments{
\item{seurat_objects}{List of Seurat objects}

\item{cell_IDs}{Adds cell identities to the beginning of each barcode}

\item{to_regress}{Variable to regress out in normalization}

\item{cluster_resolution}{Resolution for cell clustering}

\item{max_dims}{Maximum dimensions for FindNeighbors and UMAP}

\item{use_SCT}{Use SCTransform method or LogNormalize (FALSE)}

\item{sct_assay}{Assay for SCTransform normalization}

\item{save_rds_file}{Save final Seurat object}

\item{file_name}{Name of files}

\item{use_elbow_plot}{Use the ElbowPlot to determine number of componenets for FindNeighbors and UMAP}

\item{spatial}{If your data is a spatial (i.e., Visium, Xenium)}

\item{integration}{Method for integrating data, see IntegrateLayers}

\item{integration_normalization}{Normalization method used}

\item{integration_assay}{Assay that has normalized data}

\item{integration_reduction}{Reduction to use for integration}

\item{new_reduction}{Name of new integrated reduction}

\item{k_anchor}{How many neighbors (k) to use when picking anchors}

\item{k_weight}{Number of neighbors to consider when weighting anchors}

\item{markers}{Find all markers}

\item{object_names}{Names for the Seurat objects}
}
\value{
An integrated Seurat object.
}
\description{
This function merges and integrates a list of Seurat objects. By default, it
assumes the objects are SCT normalized and will be integrated using Harmony.
It can integrate ATAC objects, as well. This function also save multiple
Seurat objects along the way and runs FindAllMarkers on the clusters.
}
