% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetImageBoundary.R
\name{SetImageBoundary}
\alias{SetImageBoundary}
\title{Set Boudnary of Images}
\usage{
SetImageBoundary(seurat_obj, boundary)
}
\arguments{
\item{boundary}{Boundry (segmentation or centroids)}

\item{obj}{Seurat object (spatial)}
}
\value{
a Seurat object
}
\description{
This function changes the default boundary of a spatial Seurat object.
}
